backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
typer==0.12.3
rich==13.7.1
httpx==0.27.0
python-dotenv==1.0.1
